Refactor makefile
Add -Iinclude directives to compilation, plus clang utilities for
formatting and linting.

The C preprocessor is used on the lexer.l and parser.y files, as neither
bison and nor yacc accept -Iinclude directives, this is needed to avoid
duplicating declarations across the project headers in include/ and
lexer.l and parser.y. The dirty solution was to:
 #define _STDIO_H