Refactor translation for new syntax
Refactor the translation process, dividing it into translator, and
expression evaluation and arithmetic, the translation of an statement
iterates over it's instructions, and in a line for each instruction
outputs the mnemonic for each one, followed by the instruction
arguments.

The semantic validations are performed alongside the translation, when
one ocurrs the statement parsing is stoped, work is pending on upgrading
the error handling.